Merge accents and flavors to variants (#162)
* Merge accents and flavors to variants

The combination of accents and flavors was a bit confusing because they
weren't freely mixable. YaruAccent.red would be silently ignored for
YaruFlavor.kubuntu, for example. Given that both accents and flavors
are just different variants of the same theme, it was decided to merge
the two enums to avoid unnecessary confusion with the API.

* Replace getYaruXxxTheme() with variant getters

* Update README.md