Add text selection theme
Flutter master has introduced a `DefaultSelectionStyle` widget that is
by default provided from MaterialApp. If no TextSelectionTheme is set in
the current theme context, TextField gets the default style that ends up
using the primary color from MaterialApp's theme data. This is not well
compatible with the newly added YaruTheme widget, which is created as a
child of MaterialApp. Providing the desired TextSelectionTheme avoids
that TextField would "bypass" YaruTheme.

Fixes: #160