updated with new fonts (#324)
Added new static fonts from the new variable version we got this year. That includes:

Light
Regular
Medium
Bold

and their Italic variants.

The new fonts should ideally not break anything but improve readability. The biggest change is the slightly thinner stroke which is an improvement over the old font which was heavier than usual.